Fixing bug where a 32 bit float was used to increment voxel counts in grains which limited each grain to a maximum of 16.7 Million voxels. WE now use a 64 bit double for these values which gives 487 bits of integer range. Updated the Field Info Reader and Ensembe Info Reader to better reflect how the file is being read and initialized. Added code to initialize arrays that are created in the Field Info and Ensemble Info Reader filters to the 'Unknown' vlaues so that arrays have a known default value before reading the files

+++ Source/DREAM3DLib/StatisticsFilters/FindSizes.cpp
@@ -139,8 +139,8 @@ void FindSizes::find_sizes()
   float diameter;
   size_t numgrains = m->getNumFieldTuples();
 
-  DataArray<float>::Pointer m_GrainCounts = DataArray<float>::CreateArray(numgrains, "GrainCounts");
-  float* graincounts = m_GrainCounts->GetPointer(0);
+  DataArray<double>::Pointer m_GrainCounts = DataArray<double>::CreateArray(numgrains, "GrainCounts");
+  double* graincounts = m_GrainCounts->GetPointer(0);
 
   // Initialize every element to 0.0
   for (size_t i = 0; i < numgrains * 1; i++)
@@ -153,7 +153,7 @@ void FindSizes::find_sizes()
     graincounts[gnum]++;
   }
   float res_scalar = m->getXRes() * m->getYRes() * m->getZRes();
-  float vol_term = static_cast<float>( (4.0/3.0)*m_pi );
+  float vol_term = static_cast<double>( (4.0/3.0)*m_pi );
   for (size_t i = 1; i < numgrains; i++)
   {
     m_NumCells[i] = static_cast<int32_t>( graincounts[i] );
@@ -163,6 +163,7 @@ void FindSizes::find_sizes()
     m_EquivalentDiameters[i] = diameter;
   }
 }
+
 void FindSizes::find_sizes2D()
 {
   VoxelDataContainer* m = getVoxelDataContainer();
